+++ apps/workspace/eslint.config.js
@@ -1,0 +1,2 @@
+import config from '@ando/eslint-config/next';
+export default config;